fix: Convert Cart page to Cart modal
