+++ src/components/Cart/Cart.css
@@ -1,0 +1,74 @@
+.cart-content {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    max-height: 80vh;
+    align-items: center;
+}
+
+.cart-container {
+    display: flex;
+    flex-direction: column;
+    align-items: start;
+    overflow-y: scroll;
+}
+
+.cart-content a {
+  text-decoration: none;
+  color: white;
+  background-color: var(--primary-color);
+  align-self: center;
+  padding: 10px 30px;
+  border-radius: 10px;
+  font-weight: 600;
+  font-size: 1rem;
+  margin-top: 40px;
+  text-align: center;
+}
+
+.cart-content a:hover {
+  background-color: black;
+  cursor: pointer;
+}
+  
+  div .explore-btn {
+    text-align: center;
+    font-weight: 800;
+    border: none;
+    border-radius: 20px;
+    font-size: 1rem;
+    color: white;
+    background-color: var(--primary-color);
+    padding: 0.5rem 2rem;
+  }
+  .no-items-container {
+    text-align: center;
+  }
+  
+  .no-items-container button {
+    cursor: pointer;
+    background-color: black;
+  }
+
+  .payment-section {
+    display: flex;
+    flex-direction: column;
+    justify-content: start;
+    align-items: stretch;
+    margin: 0 10px;
+    width: calc(100% - 20px);
+  }
+  
+  @media (min-width: 992px) {
+    .cart-container {
+      flex-direction: row;
+      overflow-y: unset;
+      align-items: start;
+    }
+
+    .payment-section {
+        margin: 10px 0;
+        width: unset;
+    }
+  }
+
