+++ src/components/CartAmountSummary/CartAmountSummary.css
@@ -2,7 +2,7 @@
   border-radius: 10px;
   padding: 20px;
   display: flex;
-  flex-flow: column nowrap;
+  flex-flow: column wrap;
   margin: 10px 0 0 0;
   background-color: rgb(242, 242, 242);
 }
@@ -17,6 +17,7 @@
   font-weight: 600;
   font-size: 1rem;
   margin-top: 40px;
+  text-align: center;
 }
 
 .cart-price-container a:hover {
@@ -26,11 +27,12 @@
 
 .cart-price-container .total {
   font-weight: 700;
-  font-size: 35px;
+  font-size: 1rem;
   color: black;
 }
 .cart-price-container h1 {
   color: black;
+  font-size: 1.5rem;
 }
 
 .cart-price-container > * {
@@ -44,15 +46,13 @@
   color: rgb(55, 134, 18);
 }
 
-@media (min-width: 768px) {
+@media (min-width: 992px) {
   .cart-price-container {
-    width: 50vw;
-    flex-flow: column nowrap;
+    flex-flow: column wrap;
     margin: 0 0 0 15px;
     align-self: flex-start;
     border-radius: 10px;
     padding: 25px;
     display: flex;
-    flex-flow: column nowrap;
   }
 }
