+++ src/components/CartListing/CartListing.css
@@ -1,0 +1,173 @@
+.cart-product-card {
+  display: flex;
+  box-shadow: 0px 0px 31px -3px rgba(232, 232, 232, 0.75);
+  border-radius: 10px;
+  padding: 10px;
+  position: relative;
+  flex-flow: row;
+  align-items: center;
+  margin: 10px;
+  min-width: 60px;
+}
+
+.cart-img {
+  width: 180px;
+  object-fit: cover;
+  height: 120px;
+}
+
+.cart-img-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.cart-products-container {
+  align-items: start;
+  line-height: 15px;
+  font-size: smaller;
+  justify-content: start;
+  min-width: 200px;
+  width: stretch;
+}
+
+.product-actions {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  margin: 10px;
+}
+
+.product-description {
+  display: flex;
+  flex-direction: column;
+  align-items: start;
+  line-height: 15px;
+  margin-right: 5px;
+}
+
+.product-description h3 {
+  margin-bottom: 5px;
+}
+
+.product-description :not(:first-child) {
+  font-size: 0.8rem;
+  color: grey;
+}
+
+.price {
+  font-size: 1.4rem;
+  font-weight: 600;
+  align-items: center;
+  display: flex;
+  flex-direction: row;
+  align-self: start;
+  margin: 0px 0px 0px 5px;
+}
+
+.price > svg {
+  width: 1.4rem;
+  height: 1.4rem;
+  margin-right: 5px;
+}
+
+.button-section {
+  height: 70px;
+  display: flex;
+  flex-flow: column nowrap;
+  justify-content: space-around;
+  align-items: flex-start;
+  width: 100%;
+  margin-left: 10px;
+}
+
+.button-section span {
+  padding: 3px;
+  font-size: 1rem;
+  align-self: center;
+}
+
+.button-section .count-btn-container {
+  display: flex;
+  font-size: 20px;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.count-btn-container {
+  width: 60%;
+}
+
+.remove-button {
+  position: relative;
+}
+
+.remove-button > * {
+  color: var(--primary-color);
+}
+.remove-button > *:hover{
+  color: black;
+  cursor: pointer;
+}
+
+.counter-btn {
+  font-size: 17px;
+  border: none;
+  padding: 5px 15px;
+  border-radius: 10px;  
+  background-color: rgb(255, 255, 255);
+  border: 1px solid rgb(225, 225, 225);
+}
+
+.counter-btn:hover{
+  cursor: pointer;
+  color: white;
+  background-color: black;
+  border: 1px solid black;
+}
+
+@media (min-width: 992px) {
+  .cart-products-container {
+    max-width: 60%;
+    max-height: 60vh;
+    overflow-y: scroll;
+  }
+}
+
+@media (max-width: 634px) {
+  .cart-product-card {
+    border-radius: 10px;
+    position: relative;
+    flex-flow: column;
+    align-items: center;
+  }
+
+  .price {
+    margin: 10px 0 0 0;
+  }
+
+  .count-btn-container {
+    width: 80%;
+    align-self: center;
+  }
+
+  .button-section {
+    margin-left: 0px;
+  }
+}
+
+@media (max-width: 370px) {
+  .product-actions {
+    flex-direction: column;
+  }
+
+  .cart-img {
+    width: calc(100% - 20px);
+    height: calc(100% / 3 - 20px);
+  }
+
+  .button-section {
+    margin-left: 0px;
+  }
+}
