+++ src/components/Coupons/Coupons.css
@@ -43,7 +43,7 @@
   cursor: pointer;
 }
 
-@media (min-width: 768px) {
+@media (min-width: 992px) {
   .coupons-section {
     margin-bottom: 10px;
     margin-left: 15px;
